+++ frontend/components/Game.jsx
