+++ frontend/index.html
@@ -1,107 +1,13 @@
 <!doctype html>
 <html lang="en">
-    <head>
-        <meta charset="utf-8" />
-        <title>My first three.js app</title>
-        <style>
-            body {
-                margin: 0;
-            }
-            #broadcast {
-                list-style-type: none;
-                margin: 10px;
-                padding: 10px;
-                background: rgba(0, 0, 0, 0.7);
-                color: white;
-                border-radius: 5px;
-                font-family: Arial, sans-serif;
-                font-size: 14px;
-                max-width: 300px;
-                position: absolute;
-                bottom: 35px;
-                left: 10px;
-            }
-            #form {
-                padding: 0.25rem;
-                position: absolute;
-                bottom: 0;
-                left: 15px;
-                right: 0;
-                display: flex;
-                height: 3rem;
-                max-width: 300px;
-                box-sizing: border-box;
-                border-radius: 5px;
-                font-family: Arial, sans-serif;
-                font-size: 14px;
-            }
-            
-            #form input, #form button {
-                background: rgba(0, 0, 0, 0.7);
-                color: white;
-                border: 1px solid rgba(144, 144, 144, 0.7);
-                padding: 5px;
-                border-radius: 5px;
-                font-size: 14px;
-            }
-        </style>
-    </head>
-    <body>
-        <ul id="broadcast"></ul>
-        <form id="form" action="">
-            <input id="input" autocomplete="off" /><button>Send</button>
-        </form>
-        <script src="/socket.io/socket.io.js"></script>
-        <script type="module" src="/main.js"></script>
-        <script>
-            // Websocket scripts
-            const socket = io();
-            const broadcast = document.getElementById("broadcast");
-            const form = document.getElementById("form");
-            const input = document.getElementById("input");
-
-            form.addEventListener("submit", (e) => {
-                e.preventDefault();
-                console.log("event listener");
-                if (input.value) {
-                    console.log("input received");
-                    socket.emit("chat message", input.value);
-                    input.value = "";
-                }
-            });
-
-            socket.on("chat message", (msg) => {
-                const item = document.createElement("li");
-                item.textContent = msg;
-                broadcast.appendChild(item);
-                msgTimeout(item);
-                window.scrollTo(0, document.body.scrollHeight);
-            });
-
-            // User connected
-            socket.on("connected", () => {
-                const item = document.createElement("li");
-                item.textContent = "A user connected";
-                broadcast.appendChild(item);
-                msgTimeout(item);
-                window.scrollTo(0, document.body.scrollHeight);
-            });
-
-            // User disconnected
-            socket.on("disconnected", () => {
-                const item = document.createElement("li");
-                item.textContent = "A user disconnected";
-                broadcast.appendChild(item);
-                msgTimeout(item);
-                window.scrollTo(0, document.body.scrollHeight);
-            });
-
-            // Timeout for broadcasts
-            function msgTimeout(item) {
-                setTimeout(() => {
-                    broadcast.removeChild(item);
-                }, 5000);
-            }
-        </script>
-    </body>
+  <head>
+    <meta charset="UTF-8" />
+<!--    <link rel="icon" type="image/svg+xml" href="/vite.svg" />-->
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Google Earth Kart</title>
+  </head>
+  <body>
+    <div id="root"></div>
+    <script type="module" src="/src/main.jsx"></script>
+  </body>
 </html>
