+++ frontend/public/Beetle.glb
