+++ package.json
@@ -8,11 +8,24 @@
         "concurrently": "^9.1.2",
         "express": "^4.21.2",
         "socket.io": "^4.8.1",
-        "three": "^0.172.0"
+        "three": "^0.172.0",
+        "@react-three/cannon": "^6.6.0",
+        "@react-three/drei": "^9.121.4",
+        "react": "^18.3.1",
+        "react-dom": "^18.3.1"
     },
     "devDependencies": {
         "prettier": "3.4.2",
-        "vite": "^6.0.7"
+        "vite": "^6.0.7",
+        "@eslint/js": "^9.19.0",
+        "@types/react": "^19.0.8",
+        "@types/react-dom": "^19.0.3",
+        "@vitejs/plugin-react": "^4.3.4",
+        "eslint": "^9.19.0",
+        "eslint-plugin-react": "^7.37.4",
+        "eslint-plugin-react-hooks": "^5.0.0",
+        "eslint-plugin-react-refresh": "^0.4.18",
+        "globals": "^15.14.0"
     },
     "scripts": {
         "dev": "concurrently \"npm run dev:frontend\" \"npm run dev:backend\"",
